requirements.txt
imbalanced-learn==0.12.4
imblearn==0.0
importlib_resources==6.4.5
ipython==8.12.3
joblib==1.4.2
jupyterlab_pygments==0.3.0
kagglehub==0.3.4
MarkupSafe==3.0.2
matplotlib==3.9.3
matplotlib-inline
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
numpy==2.0.2
pandas==2.2.3
pillow==11.0.0
pipreqs==0.5.0
pyparsing==3.2.0
pywin32==307
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
urllib3==2.2.3
xgboost==2.1.3
